An offboarding script for sbs2011
THIS SCRIPT WILL ONLY WORK AS INTENDED ON SBS2011

This script is meant for users who are not well versed in Powershell, or
any other command line. It allows a non-technical person to do a basic
offboard of a user's account, by converting their mailbox to a Shared
Mailbox to allow for better administration through the GUI, which also
results in the user becoming disabled, preventing them from logging in.

This script was written for Powershell 2.0 as it ships on sbs2011